requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
flask==2.3.2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.0
sqlalchemy==2.0.19
apscheduler==3.10.1
pytest==7.4.0
